pyproject.toml
[tool.poetry]
name = "advanced-rag-flow"
version = "0.1.0"
description = ""
authors = ["Mario Toribio <yuniormtr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.3"
langchain = "^0.3.7"
langchainhub = "^0.1.21"
langgraph = "^0.2.45"
langchain-community = "^0.3.5"
langchain-ollama = "^0.2.0"
langchain-openai = "^0.2.6"
tavily-python = "^0.5.0"
langchain-chroma = "^0.1.4"
python-dotenv = "^1.0.1"
ruff = "^0.7.3"
isort = "^5.13.2"
pytest = "^8.3.3"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
